+++ boreas/alivedetection.c
@@ -137,19 +137,6 @@ struct hosts_data
   GHashTable *alivehosts_not_to_be_sent_to_openvas;
 };
 
-struct arp_hdr
-{
-  uint16_t htype;
-  uint16_t ptype;
-  uint8_t hlen;
-  uint8_t plen;
-  uint16_t opcode;
-  uint8_t sender_mac[6];
-  uint8_t sender_ip[4];
-  uint8_t target_mac[6];
-  uint8_t target_ip[4];
-};
-
 struct sniff_ethernet
 {
   u_char ether_dhost[ETHER_ADDR_LEN]; /* Destination host address */
@@ -650,59 +637,6 @@ print_host_str (gpointer key, __attribute__ ((unused)) gpointer value,
 }
 
 /**
- * @brief Get the source mac address of the given interface
- * or of the first non lo interface.
- *
- * @param interface Interface to get mac address from or NULL if first non lo
- * interface should be used.
- * @param[out]  mac Location where to store mac address.
- *
- * @return 0 on success, -1 on error.
- */
-static int
-get_source_mac_addr (gchar *interface, uint8_t *mac)
-{
-  struct ifaddrs *ifaddr = NULL;
-  struct ifaddrs *ifa = NULL;
-  int interface_provided = 0;
-
-  if (interface)
-    interface_provided = 1;
-
-  if (getifaddrs (&ifaddr) == -1)
-    {
-      g_debug ("%s: getifaddr failed: %s", __func__, strerror (errno));
-      return -1;
-    }
-  else
-    {
-      for (ifa = ifaddr; ifa != NULL; ifa = ifa->ifa_next)
-        {
-          if ((ifa->ifa_addr) && (ifa->ifa_addr->sa_family == AF_PACKET)
-              && !(ifa->ifa_flags & (IFF_LOOPBACK)))
-            {
-              if (interface_provided)
-                {
-                  if (g_strcmp0 (interface, ifa->ifa_name) == 0)
-                    {
-                      struct sockaddr_ll *s =
-                        (struct sockaddr_ll *) ifa->ifa_addr;
-                      memcpy (mac, s->sll_addr, 6 * sizeof (uint8_t));
-                    }
-                }
-              else
-                {
-                  struct sockaddr_ll *s = (struct sockaddr_ll *) ifa->ifa_addr;
-                  memcpy (mac, s->sll_addr, 6 * sizeof (uint8_t));
-                }
-            }
-        }
-      freeifaddrs (ifaddr);
-    }
-  return 0;
-}
-
-/**
  * @brief Is called in g_hash_table_foreach(). Check if ipv6 or ipv4, get
  * correct socket and start appropriate ping function.
  *
@@ -1102,120 +1036,6 @@ send_tcp (__attribute__ ((unused)) gpointer key, gpointer value,
       dst4.s_addr = dst6_p->s6_addr32[3];
       send_tcp_v4 (scanner.tcpv4soc, dst4_p, scanner.tcp_flag);
     }
-}
-
-/**
- * @brief Send arp ping.
- *
- * @param soc Socket to use for sending.
- * @param dst Destination address to send to.
- */
-static void
-send_arp_v4 (int soc, struct in_addr *dst_p)
-{
-  struct sockaddr_ll soca;
-  struct arp_hdr arphdr;
-  int frame_length;
-  uint8_t *ether_frame;
-
-  static gboolean first_time_setup_done = FALSE;
-  static struct in_addr src;
-  static int ifaceindex;
-  static uint8_t src_mac[6];
-  static uint8_t dst_mac[6];
-
-  memset (&soca, 0, sizeof (soca));
-
-  /* Set up data which does not change between function calls. */
-  if (!first_time_setup_done)
-    {
-      struct sockaddr_storage storage_src;
-      struct sockaddr_storage storage_dst;
-      struct sockaddr_in sin_src;
-      struct sockaddr_in sin_dst;
-
-      memset (&sin_src, 0, sizeof (struct sockaddr_in));
-      memset (&sin_dst, 0, sizeof (struct sockaddr_in));
-      sin_src.sin_family = AF_INET;
-      sin_dst.sin_family = AF_INET;
-      sin_dst.sin_addr = *dst_p;
-      memcpy (&storage_dst, &sin_dst, sizeof (sin_dst));
-      memcpy (&storage_dst, &sin_src, sizeof (sin_src));
-
-      /* Get interface and set src addr. */
-      gchar *interface = gvm_routethrough (&storage_dst, &storage_src);
-      memcpy (&src, &((struct sockaddr_in *) (&storage_src))->sin_addr,
-              sizeof (struct in_addr));
-      g_warning ("%s: %s", __func__, inet_ntoa (src));
-
-      if (!interface)
-        g_warning ("%s: no appropriate interface was found", __func__);
-      g_debug ("%s: interface to use: %s", __func__, interface);
-
-      /* Get interface index for sockaddr_ll. */
-      if ((ifaceindex = if_nametoindex (interface)) == 0)
-        g_warning ("%s: if_nametoindex: %s", __func__, strerror (errno));
-
-      /* Set MAC addresses. */
-      memset (src_mac, 0, 6 * sizeof (uint8_t));
-      memset (dst_mac, 0xff, 6 * sizeof (uint8_t));
-      if (get_source_mac_addr (interface, (unsigned char *) src_mac) != 0)
-        g_warning ("%s: get_source_mac_addr() returned error", __func__);
-
-      g_debug ("%s: Source MAC address: %02x:%02x:%02x:%02x:%02x:%02x",
-               __func__, src_mac[0], src_mac[1], src_mac[2], src_mac[3],
-               src_mac[4], src_mac[5]);
-      g_debug ("%s: Destination mac address: %02x:%02x:%02x:%02x:%02x:%02x",
-               __func__, dst_mac[0], dst_mac[1], dst_mac[2], dst_mac[3],
-               dst_mac[4], dst_mac[5]);
-
-      first_time_setup_done = TRUE;
-    }
-
-  /* Fill in sockaddr_ll.*/
-  soca.sll_ifindex = ifaceindex;
-  soca.sll_family = AF_PACKET;
-  memcpy (soca.sll_addr, src_mac, 6 * sizeof (uint8_t));
-  soca.sll_halen = 6;
-
-  /* Fill ARP header.*/
-  /* IP addresses. */
-  memcpy (&arphdr.target_ip, dst_p, 4 * sizeof (uint8_t));
-  memcpy (&arphdr.sender_ip, &src, 4 * sizeof (uint8_t));
-  /* Hardware type ethernet.
-   * Protocol type IP.
-   * Hardware address length is MAC address length.
-   * Protocol address length is length of IPv4.
-   * OpCode is ARP request. */
-  arphdr.htype = htons (1);
-  arphdr.ptype = htons (ETH_P_IP);
-  arphdr.hlen = 6;
-  arphdr.plen = 4;
-  arphdr.opcode = htons (1);
-  memcpy (&arphdr.sender_mac, src_mac, 6 * sizeof (uint8_t));
-  memset (&arphdr.target_mac, 0, 6 * sizeof (uint8_t));
-
-  /* Ethernet frame to send. */
-  ether_frame = g_malloc0 (IP_MAXPACKET);
-  /* (MAC + MAC + ethernet type + ARP_HDRLEN) */
-  frame_length = 6 + 6 + 2 + 28;
-
-  memcpy (ether_frame, dst_mac, 6 * sizeof (uint8_t));
-  memcpy (ether_frame + 6, src_mac, 6 * sizeof (uint8_t));
-  /* ethernet type code */
-  ether_frame[12] = ETH_P_ARP / 256;
-  ether_frame[13] = ETH_P_ARP % 256;
-  /* ARP header.  ETH_HDRLEN = 14, ARP_HDRLEN = 28 */
-  memcpy (ether_frame + 14, &arphdr, 28 * sizeof (uint8_t));
-
-  if ((sendto (soc, ether_frame, frame_length, MSG_NOSIGNAL,
-               (struct sockaddr *) &soca, sizeof (soca)))
-      <= 0)
-    g_warning ("%s: sendto(): %s", __func__, strerror (errno));
-
-  g_free (ether_frame);
-
-  return;
 }
 
 /**
